+++ CHANGELOG.md
@@ -102,7 +102,8 @@
 
 
 
-
+* Cloud Access
+  * Added support for the new authentication method `G2O` (Akamai Signature Header Authentication) in the `akamai_cloudaccess_key` resource.